Move the WinPcap extensions into pcap/pcap.h.
That's less complicated than putting them in a separate header.

We also get rid of all checks for HAVE_REMOTE in public headers - using
configuration #defines in public header files is The Wrong Thing To Do,
because there's no guarantee that a program using those headers will
define those particular values.
